requirements.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dialogflow==1.1.0
Flask==2.0.1
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
twilio==6.61.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
yapf==0.31.0
